Make output of seq2seq_f logit sized
This is how it is done in the embeding_attention_seq2seq model
given by tensorflow and it is assumed by the model_with_buckets
that the output from the seq2seq function has the same size
as the number of symbols in the target vocab.

On a model level this has some implication that we might have
thought about. We are not modelling the probability of the next word
one by one but actully the whole distribution as anticipated...